2025 06 16 add daq,start condition check
daq,start now will check if the consumer for SBE41 would stop due to OS problems, if so, it will return nak within 10 seconds of daq,start request
mod_som_main task also monitors all tasks started by daq,start and if any task failed, it would resume the task

+++ mod/efe/mod_som_efe.h
@@ -248,7 +248,7 @@ typedef struct{
   uint32_t  max_sample_per_record;   //ALB maximum element per stream block
   uint32_t  payload_length;
   uint32_t  elmnts_skipped;
-  uint64_t  cnsmr_cnt;
+  uint32_t  cnsmr_cnt;
 
   uint8_t   data_ready_flg;           //ALB ???
   uint8_t   chksum;
@@ -289,7 +289,7 @@ mod_som_efe_data_consumer_t, *mod_som_efe_data_consumer_ptr_t;
 typedef struct{
     uint32_t initialized_flag;
     uint32_t error_flag;
-    uint64_t sample_count;
+    uint32_t sample_count;
     uint64_t timestamp;
     mod_som_efe_settings_ptr_t settings_ptr;      //
     mod_som_efe_config_ptr_t config_ptr;    //
